requirements.txt
# Web framework
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
python-multipart>=0.0.9,<0.1

# Database
sqlalchemy>=2.0.30,<2.1
asyncpg>=0.29,<0.30
alembic>=1.13,<1.14
greenlet>=3.0,<4.0

# Task queue
celery[redis]>=5.4,<5.5
redis>=5.0,<5.1
celery-redbeat>=2.2,<2.4

# Auth & security
python-jose[cryptography]>=3.3,<3.4
passlib[bcrypt]>=1.7,<1.8
bcrypt>=3.2,<4.0
cryptography>=42.0,<43.0

# Config
pydantic[email]>=2.7,<3.0
pydantic-settings>=2.3,<3.0
python-dotenv>=1.0,<2.0

# HTTP client
httpx>=0.27,<0.28

# Logging
loguru>=0.7,<0.8

# Alembic sync driver (used only by alembic env.py, never by the app)
psycopg2-binary>=2.9,<3.0

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<0.24
pytest-cov>=5.0,<6.0
respx>=0.21,<0.22
